+++ test/rgb2hsi.test.js
@@ -3,17 +3,21 @@ const assert = require('assert');
 require('es6-shim');
 
 const rgb2hsi = require('../src/io/hsi/rgb2hsi');
+const chroma = require('../src/chroma');
 
 const tests = {
-    black2:     { hsi: [NaN,0,0],   rgb: [0,0,0,1]},
-    white:      { hsi: [NaN,0,1],   rgb: [255,255,255,1]},
-    gray:       { hsi: [NaN,0,0.5],   rgb: [127.5,127.5,127.5,1]},
-    red:        { hsi: [0,1,1/3],   rgb: [255,0,0,1]},
-    yellow:     { hsi: [60,1,2/3],  rgb: [255,255,0,1]},
-    green:      { hsi: [120,1,1/3], rgb: [0,255,0,1]},
-    cyan:       { hsi: [180,1,2/3], rgb: [0,255,255,1]},
-    blue:       { hsi: [240,1,1/3], rgb: [0,0,255,1]},
-    magenta:    { hsi: [300,1,2/3], rgb: [255,0,255,1]},
+    black2:       { hsi: [NaN,0,0],   rgb: [0,0,0,1]},
+    white:        { hsi: [NaN,0,1],   rgb: [255,255,255,1]},
+    gray:         { hsi: [NaN,0,0.5], rgb: [127.5,127.5,127.5,1]},
+    black2:       { hsi: [0,0,0],     rgb: [0,0,0,1],             noHueAsZero: true},
+    whiteZeroHue: { hsi: [0,0,1],     rgb: [255,255,255,1],       noHueAsZero: true},
+    grayZeroHue:  { hsi: [0,0,0.5],   rgb: [127.5,127.5,127.5,1], noHueAsZero: true},
+    redZeroHue:   { hsi: [0,1,1/3],   rgb: [255,0,0,1]},
+    yellow:       { hsi: [60,1,2/3],  rgb: [255,255,0,1]},
+    green:        { hsi: [120,1,1/3], rgb: [0,255,0,1]},
+    cyan:         { hsi: [180,1,2/3], rgb: [0,255,255,1]},
+    blue:         { hsi: [240,1,1/3], rgb: [0,0,255,1]},
+    magenta:      { hsi: [300,1,2/3], rgb: [255,0,255,1]},
 };
 
 const round = (digits) => {
@@ -28,14 +32,17 @@ Object.keys(tests).forEach(key => {
     batch[`rgb2hsi ${key}`] = {
         topic: tests[key],
         array(topic) {
+            chroma.noHueAsZero(topic.noHueAsZero || false);
             assert.deepStrictEqual(rgb2hsi(topic.rgb).map(rnd), topic.hsi.map(rnd));
         },
         obj(topic) {
             let [r,g,b] = topic.rgb;
+            chroma.noHueAsZero(topic.noHueAsZero || false);
             assert.deepStrictEqual(rgb2hsi({r,g,b}).map(rnd), topic.hsi.map(rnd));
         },
         args(topic) {
             if (topic.mode != 'auto') return
+            chroma.noHueAsZero(topic.noHueAsZero || false);
             assert.deepStrictEqual(rgb2hsi.apply(null, topic.rgb).map(rnd), topic.hsi.map(rnd));
         }
     }